Update spec for certora-cli 8.x compatibility...
The keyword "STORAGE" is no longer allowed when defining storage hooks.

Removing this, however, makes it incompatible with versions of
certora-cli before 7.0.7. Another solution could be to instruct students
to use a specific version (e.g. 7.6.3, which is shown in the written
lesson notes for video 6) ... though they will see a CLI incompatibility
warning because it is not the latest.